DashAttackState, IgnoreEntities, NoEnemyZone, AbstractEnemyAbilityState
### Added
- [AbstractEnemyAbilityState] Abstraction for states that have the common logic of an ability, cooldowns, canCast() etc.
- [DashAttackState] Dash attack state fully implemented
- [IgnoreEntities, NoEnemyZone layers] IgnoreEntites doesnt collide with itself or the layer Entities. NoEnemyZone blocks all enemy vision as well as A*.
- [OrcDasherCharging animation] Animation for OrcDasher precharge before dashing

+++ Assets/Scripts/Entities/IDamageable.cs
@@ -1,4 +1,4 @@
 ﻿public interface IDamageable 
 {
-    void Damage(int amount);
+    void Damage(float amount);
 }
